requirements.txt
absl-py==2.0.0
antlr4-python3-runtime==4.9.3
anyio==4.2.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.2
bleach==6.1.0
brotlipy==0.7.0
cachetools==5.3.2
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.13.1
fonttools==4.47.0
fqdn==1.5.1
fsspec==2023.12.2
google-auth==2.26.1
google-auth-oauthlib==1.2.0
grpcio==1.60.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hydra-core==1.3.2
iopath==0.1.10
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
json5==0.9.14
jsonpointer==2.4
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.1
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.2
jupyter_server_terminals==0.5.1
jupyterlab==4.2.5
jupyterlab-language-pack-zh-CN==4.0.post6
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.5
Markdown==3.5.1
MarkupSafe==2.1.3
matplotlib==3.9.2
matplotlib-inline==0.1.6
mistune==3.0.2
mpmath==1.3.0
nbclient==0.9.0
nbconvert==7.14.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
notebook==7.2.1
notebook_shim==0.2.3
numpy==1.26.3
oauthlib==3.2.2
omegaconf==2.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
overrides==7.4.0
packaging==23.2
pandas==2.2.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
portalocker==2.10.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.2
QtPy==2.4.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.16.2
rsa==4.9
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
supervisor==4.2.5
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
#torch==2.3.1+cu121
#torchaudio==2.3.1+cu121
#torchvision==0.18.1+cu121
tornado==6.4
tqdm==4.66.5
traitlets==5.14.1
#triton==2.3.1
typing_extensions==4.9.0
tzdata==2024.1
ultralytics==8.2.82
ultralytics-thop==2.0.5
uri-template==1.3.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
Werkzeug==3.0.1
widgetsnbextension==4.0.9
